==> 4_Playground.ipynb <==
import pandas as pd

tbbt_ngramsdf = pd.read_csv('FINAL_TBBT_ngrams.csv')
himym_ngramsdf = pd.read_csv('FINAL_HIMYM_ngrams.csv')
friends_ngramsdf = pd.read_csv('FINAL_Friends_ngrams.csv')

merged_df = pd.merge(tbbt_ngramsdf[['ngram', 'ngram_type', 'count', 'Sheldon', 'Leonard', 'Penny', 'Howard', 'Raj', 'Bernadette', 'Amy']], 
                     himym_ngramsdf[['ngram', 'ngram_type', 'count', 'Barney', 'Ted', 'Lily', 'Robin', 'Marshall']],
                     on='ngram', 
                     how='outer', 
                     suffixes=('_TBBT', '_HIMYM'))


merged_df = pd.merge(merged_df, 
                     friends_ngramsdf[['ngram', 'ngram_type', 'count', 'Monica', 'Joey', 'Chandler', 'Phoebe', 'Ross', 'Rachel']], 
                     on='ngram', 
                     how='outer')

merged_df = merged_df.rename(columns={'count': 'Count_Friends'})

merged_df['ngram_type'] = merged_df[['ngram_type_TBBT', 'ngram_type_HIMYM', 'ngram_type']].bfill(axis=1).iloc[:, 0]

merged_df = merged_df.drop(['ngram_type_TBBT', 'ngram_type_HIMYM'], axis=1)

merged_df.insert(1, 'ngram_type', merged_df.pop('ngram_type'))

merged_df['Count_ALL'] = merged_df[['count_TBBT', 'count_HIMYM', 'Count_Friends']].sum(axis=1, skipna=True)

merged_df.to_csv('FINAL_merged_ngrams.csv', index=False)
